+++ src/core/QueryManager.ts
@@ -1101,7 +1101,7 @@ export class QueryManager {
           // return a chainable promise
           this.removeFetchQueryPromise(requestId);
           resolve({ data: resultFromStore, loading: false, networkStatus: NetworkStatus.ready, stale: false });
-          return null;
+          return Promise.resolve();
         }).catch((error: Error) => {
           reject(error);
         });