+++ test/readFromStore.ts
@@ -8,6 +8,9 @@ import {
 import {
   NormalizedCache,
   StoreObject,
+  IdValue,
+  StoreValue,
+  JsonValue,
 } from '../src/data/storeUtils';
 
 import {
@@ -170,7 +173,7 @@ describe('reading from the store', () => {
           id: 'abcde',
           generated: false,
         },
-      }) as StoreObject,
+      } as StoreObject),
       abcde: result.nestedObj,
     } as NormalizedCache;
 
@@ -228,7 +231,7 @@ describe('reading from the store', () => {
           id: 'abcde',
           generated: false,
         },
-      }) as StoreObject,
+      } as StoreObject),
       abcde: assign({}, result.nestedObj, {
         deepNestedObj: {
           type: 'id',
@@ -362,8 +365,8 @@ describe('reading from the store', () => {
     const store = {
       'ROOT_QUERY': assign({}, assign({}, omit(result, 'nestedArray')), {
         nestedArray: [
-          { type: 'id', generated: true, id: 'abcd.nestedArray.0' },
-          { type: 'id', generated: true, id: 'abcd.nestedArray.1' },
+          { type: 'id', generated: true, id: 'abcd.nestedArray.0' } as IdValue,
+          { type: 'id', generated: true, id: 'abcd.nestedArray.1' } as IdValue,
         ],
       }) as StoreObject,
       'abcd.nestedArray.0': result.nestedArray[0],
@@ -421,7 +424,7 @@ describe('reading from the store', () => {
       'ROOT_QUERY': assign({}, assign({}, omit(result, 'nestedArray')), {
         nestedArray: [
           null,
-          { type: 'id', generated: true, id: 'abcd.nestedArray.1' },
+          { type: 'id', generated: true, id: 'abcd.nestedArray.1' } as IdValue,
         ],
       }) as StoreObject,
       'abcd.nestedArray.1': result.nestedArray[1],
@@ -583,7 +586,7 @@ describe('reading from the store', () => {
       'ROOT_QUERY': assign({}, assign({}, omit(result, 'simpleArray')), { simpleArray: {
         type: 'json',
         json: result.simpleArray,
-      }}) as StoreObject,
+      } as JsonValue }) as StoreObject,
     } as NormalizedCache;
 
     const queryResult = readQueryFromStore({
@@ -618,7 +621,7 @@ describe('reading from the store', () => {
       'ROOT_QUERY': assign({}, assign({}, omit(result, 'simpleArray')), { simpleArray: {
         type: 'json',
         json: result.simpleArray,
-      }}) as StoreObject,
+      } as JsonValue }) as StoreObject,
     } as NormalizedCache;
 
     const queryResult = readQueryFromStore({
@@ -745,7 +748,7 @@ describe('reading from the store', () => {
           type: 'id',
           id: 'abcde',
           generated: false,
-        },
+        } as IdValue,
       }) as StoreObject,
       abcde: assign({}, data.nestedObj, {
         deepNestedObj: {